add local and real tests
