+++ build.gradle
@@ -1,6 +1,6 @@
 plugins {
     id 'java-library'
-    id 'io.qameta.allure' version '2.10.0'
+    id 'io.qameta.allure' version '2.11.0'
     id "io.freefair.lombok" version "6.0.0-m2"
 }
