--- src/main/java/tests/TestBase.java
@@ -1,39 +1,0 @@
-package tests;
-
-import com.codeborne.selenide.Configuration;
-import drivers.BrowserstackMobileDriver;
-import helpers.Attach;
-import io.qameta.allure.selenide.AllureSelenide;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-
-import static com.codeborne.selenide.Selenide.*;
-import static com.codeborne.selenide.logevents.SelenideLogger.addListener;
-
-public class TestBase {
-    @BeforeAll
-    public static void setup() {
-        Configuration.browser = BrowserstackMobileDriver.class.getName();
-        Configuration.browserSize = null;
-    }
-
-    @BeforeEach
-    public void startDriver() {
-        addListener("AllureSelenide", new AllureSelenide());
-
-        open();
-    }
-
-    @AfterEach
-    public void afterEach() {
-        String sessionId = sessionId().toString();
-
-        Attach.screenshotAs("Last screenshot");
-        Attach.pageSource();
-
-        closeWebDriver();
-
-        Attach.video(sessionId);
-    }
-}
