+++ src/test/java/drivers/BrowserstackMobileDriver.java
@@ -1,0 +1,48 @@
+package drivers;
+
+import com.codeborne.selenide.WebDriverProvider;
+import lombok.SneakyThrows;
+import org.aeonbits.owner.ConfigFactory;
+import org.openqa.selenium.Capabilities;
+import org.openqa.selenium.MutableCapabilities;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.remote.RemoteWebDriver;
+
+import java.net.MalformedURLException;
+import java.net.URL;
+
+public class BrowserstackMobileDriver implements WebDriverProvider {
+    private static MobileConfig config = ConfigFactory.create(MobileConfig.class,System.getProperties());
+    @SneakyThrows
+    @Override
+    public WebDriver createDriver(Capabilities capabilities) {
+        MutableCapabilities mutableCapabilities = new MutableCapabilities();
+        mutableCapabilities.merge(capabilities);
+        // Set your access credentials
+        mutableCapabilities.setCapability("browserstack.user", config.getBrowserstackUser());
+        mutableCapabilities.setCapability("browserstack.key", config.getBrowserstackKey());
+
+        // Set URL of the application under test
+        mutableCapabilities.setCapability("app", config.getApp());
+
+        // Specify device and os_version for testing
+        mutableCapabilities.setCapability("device", config.getDevice());
+        mutableCapabilities.setCapability("os_version", config.getOsVersion());
+
+        // Set other BrowserStack capabilities
+        mutableCapabilities.setCapability("project", config.getProject());
+        mutableCapabilities.setCapability("build", config.getBuild());
+        mutableCapabilities.setCapability("name", config.getName());
+
+        // Initialise the remote Webdriver using BrowserStack remote URL
+        // and desired capabilities defined above
+        return new RemoteWebDriver(new URL(config.getURL()), mutableCapabilities);
+    }
+    public URL getBrowserstackUrl() {
+        try {
+            return new URL(config.getURL());
+        } catch (MalformedURLException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
