+++ src/test/java/tests/browserstack/AndroidSearchTests.java
@@ -1,0 +1,43 @@
+//package tests.browserstack;
+//
+//import io.appium.java_client.AppiumBy;
+//import org.junit.jupiter.api.Test;
+//import tests.TestBase;
+//
+//
+//import static com.codeborne.selenide.CollectionCondition.sizeGreaterThan;
+//import static com.codeborne.selenide.Selenide.$;
+//import static com.codeborne.selenide.Selenide.$$;
+//import static groovy.xml.dom.DOMCategory.text;
+//import static io.qameta.allure.Allure.step;
+//
+//
+//public class AndroidSearchTests extends TestBase {
+//
+//    @Test
+//    void searchTest() {
+//        step("Type search", () -> {
+//            $(AppiumBy.accessibilityId("Search Wikipedia")).click();
+//            $(AppiumBy.id("org.wikipedia.alpha:id/search_src_text")).sendKeys("BrowserStack");
+//        });
+//        step("Verify content found", () ->
+//                $$(AppiumBy.id("org.wikipedia.alpha:id/page_list_item_title"))
+//                        .shouldHave(sizeGreaterThan(0)));
+//    }
+//    @Test
+//    void searchDisneylandTest() {
+//        step("Type search", () -> {
+//            $(AppiumBy.accessibilityId("Search Wikipedia")).click();
+//            $(AppiumBy.id("org.wikipedia.alpha:id/search_src_text")).sendKeys("Disneyland");
+//        });
+//
+//        step("Verify content found", () ->
+//                $$(AppiumBy.id("org.wikipedia.alpha:id/page_list_item_title"))
+//                        .shouldHave(sizeGreaterThan(0)));
+//
+//        /*$$(AppiumBy.id("org.wikipedia.alpha:id/page_list_item_title")).find(Condition.text("Disneyland")).click();
+//        $$(AppiumBy.id("org.wikipedia.alpha:id/pcs-edit-section-title-description"))
+//                .find(Condition.text("Disneyland"));*/
+//    }
+//
+//}
