+++ src/test/java/tests/local/TestBaseLocal.java
@@ -1,0 +1,38 @@
+package tests.local;
+
+import com.codeborne.selenide.Configuration;
+import com.codeborne.selenide.Selenide;
+import drivers.LocalMobileDriver;
+import helpers.Attach;
+import io.appium.java_client.AppiumBy;
+import io.qameta.allure.selenide.AllureSelenide;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+
+import static com.codeborne.selenide.Selenide.$;
+import static com.codeborne.selenide.Selenide.open;
+import static com.codeborne.selenide.logevents.SelenideLogger.addListener;
+import static io.qameta.allure.Allure.step;
+
+public class TestBaseLocal {
+    @BeforeAll
+    public static void setup() {
+        Configuration.browser = LocalMobileDriver.class.getName();
+        Configuration.browserSize = null;
+    }
+
+    @BeforeEach
+    public void startDriver() {
+        addListener("AllureSelenide", new AllureSelenide());
+
+    }
+
+    @AfterEach
+    public void afterEach() {
+        Attach.screenshotAs("Last screenshot");
+        Attach.pageSource();
+
+        step("Close driver", Selenide::closeWebDriver);
+    }
+}
